requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
chromadb
google-genai
python-dotenv
python-multipart
mlflow
cryptography
pytest
transformers
torch
streamlit
requests